# docs/swagger.yaml
basePath: /
definitions:
...
  shared.ErrorResponse:
    properties:
      code:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  shared.LegalDocumentItem:
    properties:
      id:
        type: string
      isCentralPolicy:
        type: boolean
      required:
        type: boolean
      sortOrder:
        type: integer
      title:
        type: string
      url:
        type: string
    type: object
...
  shared.RegistrationConfig:
    properties:
      active:
        type: boolean
      fieldConfig:
        additionalProperties:
          type: string
        type: object
      introText:
        type: string
      legalDocuments:
        items:
          $ref: '#/definitions/shared.LegalDocumentItem'
        type: array
      rcNumber:
        type: string
      sepaMandateEnabled:
        type: boolean
      showCentralPolicy:
        type: boolean
      title:
        type: string
    type: object
...
host: member-onboarding.eegfaktura.at
info:
  contact: {}
  description: |-
    REST API for EEG (Energiegemeinschaft) member self-service registration and admin review.

    **Auth schemes:**
    - **Public** endpoints (`/api/public/*`): no auth required. Rate-limited + optional Turnstile CAPTCHA.
    - **Admin** endpoints (`/api/admin/*`): Keycloak Bearer JWT (`Authorization: Bearer <token>`).
    - **External** endpoints (`/api/external/*`): API key (`Authorization: Bearer moak_<key>`).
  title: eegfaktura Member Onboarding API
  version: "1.0"
paths:
...
  /api/public/registration/{rc_number}:
    get:
      description: 'Returns EEG-specific configuration for the public registration
        form: field visibility, legal documents, intro text, SEPA settings.'
      parameters:
      - description: EEG registration code (e.g. RC123456)
        in: path
        name: rc_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.RegistrationConfig'
        "404":
          description: RC number unknown
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "410":
          description: Registration deactivated for this EEG
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Get registration config
      tags:
      - Public
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'EEG API key. Format: "Bearer moak_<key>"'
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    description: 'Keycloak JWT. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"